Fix compiler errors in logger.ts
Change debugWindow, logBuffer, and logScroll to use the private keyword
instead of '#' for constancy with the rest of the progect.

Add type annotations to the properties of Logger to remove implied 'any'
type errors.

Add null checks for debugWindow and the text element so that the code is
more verbose when it is unable to interface with the debugWindow DOM
correctly.