Add image tracking to Webpack
Add the ability to import images with the import statement so that
Webpack can bundle the images.

Add '*.svg' module type so that typescript recognises the image import
as valid. This must be done for all new file types that are imoprted.
For example, '*.css', '*.tff', '*.png', etc. More documentation abobut
this can be found here:
https://webpack.js.org/guides/typescript/#importing-other-assets