Remove try-catch statementts
Remove try-catch statements so that the error is thrown to the global
scope. This can then be picked up by the VSCode Debugger console. This
is better because it also works on chromebook.

Configure vscode launch.json to be able to easily start a debugging
session.

With this change, there is a bug where the source maps do not accurately
reflect the source line numbers.

Fix array indexing when creating a new TiledTexture. Previously, the
array was indexed with only the random number, which produced
non-integer indexes. Replace this with a Math.floor() call and multiply
the random number by the length of the array. This will generate a
random number between 0 and 1 - array.length.