Prepare for npm publish
Move all example files to another directory for teseting. This new
directory will be called 'merlin-game-engine-example'.

Rename 'main.ts' to 'index.ts' to match convention.
Update imports in all files that refrence 'main.ts' to refrence
'index.ts'.

Uninstall webpack, webpack-cli, webpack-dev-server, and ts-loader
because those dependencies are not needed when publishing to npm.

Remove "sourceMap: true" from tsconfig because source maps are not
needed in an npm publish.
Add "declaration: true" and specify the output directory for tsc so that
types are created for typescript users of the game engine, and so that
the compiled code is in one directory.