Setup TypeDoc and Add TsDocs for CanvasLayer
Add typedoc as a dev-dependency. Creae a docs folder where the compiled
typedoc will live. This directory will be added to .gitignore because
the docs do not have to live in source control. Additionally, it is
easy to add the docs/ directory to git later.

Create TSDocs for the CanvasLayer class and its methods and properties.