Add Tests
Add tests by using webpack. Webpack hosts everything from the
test/public directory, and finds the source in test/src.

Add a basic test copied from another project to teset collisions between
multiple kinematicBodies. Found an issue where if the regular player and
the square player are in the same location, i.e. overlapping, then the
object that calls moveAndSlide() will slowly be pushed out. This should
not happen, as it should be immediately pushed all the way out. If the
squarePlayer were to also call moveAndSlide(), then it would move
upwards with the regular player at the same time, in an infinite loop.