Fix SAT collision
Fix Seperating Axis Theorem (SAT) collision tests by changing the final
position given for the collision to be based on the sum of
the halfsizes of the two collision objects minus the distance between
the centers of the two collision objects (px and py).
signX or signY must be 1 if dx or dy is equal to 0 because it will
prevent nothing from hapening if the two objects perfectly align.