Create a different tsconfig for production
Create `tsconfig.production.json` to create the comipled source code
before distribution with `npm publish` or `npm pack`.

The `build` npm script will update to use the production tsconfig.
The `prepublish` npm script will update to `prepack` because `prepublish`
is no longer supported and does not work. `prepack` runs before a
tarball of the project is created, i.e. running `npm publish` or `npm
pack`.
`npm run build` should be run before running `npm publish` so that the
result can be checked before published.

The `html-webpack-plugin` settings in `webpack.config.js` will be
change so that the title of the generated webpage is not the default.
This is purely for asthetic reasons and does not affect the performance
in any way.

The test game will not default to being paused because it can cause a
lot of confusion if `npm test` is run, but nothing is displayed. For
this same reason, projects should avoid keeping the engine paused on
start.