Add AABB vs Segment collision test
Add a AABB vs Line Segment collision test in the physics engine. The
test finds the time of the segment's intersection with the near and far
edges of the AABB for each axis of the AABB. If they overlap, then the
segment is intersecting. A great website detailing the code and the test
can be found at https://noonat.github.io/intersect/#aabb-vs-segment.