Add collision tests from Noonat.github.io
Add a normalize method for vector2 that returns a new vector that has
been normalized. This is a generally useful operation on vectors, but
it was required in the intersection tests.

Add seperating axis theorem AABB collision as its own method on
`PhysicsEngine`. This will be called on by the new swept AABB collision
test.

Add Minkowski swept AABB collision tests. The exact details can be found
at https://noonat.github.io/intersect/#aabb-vs-swept-aabb, but the
general idea is that the velocity of one of the colliders relative to
the other can be found by subtracting the velocities. This velocity is
used as the delta of a line segment starting at the midpoint of the
first collider. The second collider is padded using the half size of the
first collider, and then a ray vs. AABB collision test is done.

Note: These new collision tests are not being used anywhere yet.