Somewhat fix KinematicBody collisions
The collision between KinematicBodies and StaticBodies will be mostly
fixed. There will still be some spacing between the objects, along with
some bouncing. This is caused by roundnning errors. This might be able
to be fixed by checking for the normal of the collision, then snapping
the correcti sides of the two objects together.

There is a bug in this version where if the squareplayer pushes the
player towards the ground, the player will be forced underground.