Refactor collision resolution in moveAndSlide()
Seperate out each resolution scenario into its own method. This allows
for easier readability, and makes it harder to miss a scanerio. When
calling a resolution, the order of the colliders may be swapped around
so that it matches the required order. This will most likely be changed
since it might cause issues where the colliders were swapped, but the
resolution program didn't know or expect it.

Write out what should happen with the new collision testing system in
the collisionTests.md file.