Add lastFrameCollisions property to RigidBodies
Add a lastFrameCollisions property to RigidBodies. The
lastFrameCollisions array is an array of CollisionData that keeps track
of each collision that happened to the object in each frame. The array
is updated with logCollision(collision) and resetFrameCollisions().
resetFrameCollisions() sets the lastFrameCollisions array to an empty
array, and logCollision(collision) pushes the collision into the
array.

There is an issue if there should be two collisions in a frame. The
physicsEngine update method will check for a collision, but if it finds
one, it won't keep checking for collisions, like
KinematicBody.moveAndSlide(). The fix for this is simply to check for
more collisions involving a given collider in one frame.

Each time the velocity of a collider is applied to the position,
multiply the velocity by delta time. This will resolve the issue where
every movement is much slower than it should, also cauing the physics
engine to find a collision before it visually seems like they should
collide.