Solve both colliders' resolutions at the same time
After detecting a collision, calculate the resolution for both Collider
A and Collider B. This is done to elimate problems where the two
colliders would resolve differently because they didn't know where the
other would go. Once the collision has been resolved, the colliders are
marked together that they already collided, so that double-counting
doesn't occur.