Fix corner collisions
Fix collisions not being found if the player is on the corner of a wall
and ground. This is done by checking the for collisions multiple times.
If a collision was found, resolve it and set the desired position and
velocity of the player to the player's variables. After there was a null
collision found, or a limit on the number of collision tests was
reached, then colliderA's position and velocity are reset to what they
were in the beginning of the update.