Fix squarePlayer collision allignment
When the squarePlayer pushes the player into a corner, colliding on the
x-axis, the squarePlayer should not leave any space inbetween the
colliders. To do this, the velocity of colliderA should be set to 0
inbetween the multiple slide check. This prevents the player from moving
too far when checking the collisions.

The y-axis for the collision normal in satAABB() must be multiplied by
-1 so that the normal corresponds to the correct side of the other
collider.