Fix squarePlayer going through ground
Fix the issue where the squarePlayer could go through the ground if the
player was standing on top of it. The bug is caused by the squarePlayer's
resolution with the player taking priority over its resolution with the
ground. This will be fixed by sorting the resolutions before applying
them. The sort order is the same as detecting the collisions, except the
determining factor if both colliders are the same type is the collision
time.