Fix squarePlayer going through player in corner
Fix the issue where the squarePlayer will go through the platformer
player if the platformerPlayer is against the wall, and the squarePlayer
squishes it.

Check to see if colliderA and colliderB have previously collided before
creating a null-collision collision resolution. This prevents a
kinematicBody from moving by its own velocity if another collider
already detected a collision with it.

Only stop a collision from being checked if the two colliders already
collided in the check immediately preceeding the current check. For
example, the player collides with the squarePlayer, which causes the
player to collide with the wall, which causes the player to push the
squarePlayer back. This would not be possible if the collision is
ignored if the colliders already collided previously, without a
collision inbetween.

Only add a collision to the final resolutions if it is the most recent
collision between the two colliders.