Change collision normal calculation
Change the collision normal calculation in rayAABB() collision tests.
The new way of calculating the normal is based on the position of the
two colliders, as opposed to the relative velocity. Having the normal
be calculated by the velocity can cause the normal to be incorrect, and
calculating based off of the positions fixes these isues.